refactor(RandomNumberGenerator): 무작위 숫자 생성 인터페이스 추가
`RandomNumberGenerator` 인터페이스와 이를 상속 `pickRandomNumber` 메서드를 통해 테스트에 용이할 수 있도록 함.
