+++ src/main/java/racingcar/model/Cars.java
@@ -1,6 +1,6 @@
 package racingcar.model;
 
-import racingcar.util.RandomNumberGenerator;
+import racingcar.util.generator.NormalRandomNumberGenerator;
 import racingcar.util.validator.InputValidator;
 
 import java.util.ArrayList;
@@ -29,8 +29,7 @@ public class Cars {
 
     public void moveForwardByRandomNumber() {
         for (Car car : cars) {
-            RandomNumberGenerator randomNumberGenerator = new RandomNumberGenerator();
-            car.canIMove(randomNumberGenerator.pickRandomNumber());
+            car.canIMove(new NormalRandomNumberGenerator());
         }
     }
