+++ src/main/java/racingcar/util/generator/NormalRandomNumberGenerator.java
@@ -1,8 +1,9 @@
-package racingcar.util;
+package racingcar.util.generator;
 
 import camp.nextstep.edu.missionutils.Randoms;
 
-public class RandomNumberGenerator {
+public class NormalRandomNumberGenerator implements RandomNumberGenerator {
+    @Override
     public int pickRandomNumber() {
         return Randoms.pickNumberInRange(0, 9);
     }
