+++ src/main/java/racingcar/util/generator/RandomNumberGenerator.java
@@ -1,0 +1,5 @@
+package racingcar.util.generator;
+
+public interface RandomNumberGenerator {
+    int pickRandomNumber();
+}
