+++ src/main/java/racingcar/util/generator/RandomNumberUnderFour.java
@@ -1,0 +1,10 @@
+package racingcar.util.generator;
+
+import camp.nextstep.edu.missionutils.Randoms;
+
+public class RandomNumberUnderFour implements RandomNumberGenerator {
+    @Override
+    public int pickRandomNumber() {
+        return Randoms.pickNumberInRange(0, 3);
+    }
+}
